fix(doctor): normalize Statfs_t inode counters to uint64 for freebsd (gu-h63v)
On FreeBSD, syscall.Statfs_t.Ffree is declared as int64 while Files is
uint64, causing a mismatched-types build failure under GOOS=freebsd:

  tmp_inodes_check_unix.go:22:14: invalid operation: total >= free
    (mismatched types uint64 and int64)
  tmp_inodes_check_unix.go:23:10: invalid operation: total - free
    (mismatched types uint64 and int64)

Cast both fields to uint64 so a single _unix build tag covers Linux,
darwin, and FreeBSD. The comment explains the FreeBSD quirk and the
defensive behavior if Ffree is ever reported as negative (the existing
total >= free guard keeps used at zero).

Verified with GOOS=freebsd GOARCH=amd64 go build ./... — previously
failed, now succeeds.

+++ internal/doctor/tmp_inodes_check_unix.go
@@ -16,8 +16,15 @@ func readTmpInodeUsage(path string) (tmpInodeUsage, error) {
 		return tmpInodeUsage{}, fmt.Errorf("statfs %s: %w", path, err)
 	}
 
-	total := st.Files
-	free := st.Ffree
+	// Normalize to uint64: on FreeBSD, Statfs_t.Ffree is declared as int64
+	// while Files is uint64. Other Unix variants (Linux, darwin) expose both
+	// as uint64. An explicit cast keeps one _unix build tag covering all of
+	// them. In the exceedingly rare case that Ffree is reported as negative
+	// (treated as "unknown" by some filesystems), the cast wraps to a very
+	// large uint64 which then falls through the total >= free guard and
+	// leaves used at zero.
+	total := uint64(st.Files)
+	free := uint64(st.Ffree)
 	var used uint64
 	if total >= free {
 		used = total - free
